+++ cli/build.js
@@ -17,6 +17,8 @@ OPTIONS:
         --config   <file>   specify the lume config file        Default: _config.js
         --location <url>    the domain for your site            Default: http://localhost
     -d, --dev               enable dev mode (view draft pages)
+        --metrics [<file>]  show the performance metrics or save them in a file
+        --verbose  <level>  different level of details (0/1/2)  Default: 1
 
     -s, --serve             start a live-reloading web server
     -p, --port     <port>   the port where the server runs      Default: 3000
@@ -25,7 +27,16 @@ OPTIONS:
 
 export async function run(args) {
   const options = parse(args, {
-    string: ["root", "src", "dest", "config", "location", "port"],
+    string: [
+      "root",
+      "src",
+      "dest",
+      "config",
+      "location",
+      "metrics",
+      "verbose",
+      "port",
+    ],
     boolean: ["dev", "serve", "open"],
     alias: { dev: "d", serve: "s", port: "p", open: "o" },
     ["--"]: true,
@@ -44,15 +55,33 @@ export async function run(args) {
 
   const site = await buildSite(options);
   console.log();
-  await site.build();
+  await site.build(options.serve);
   console.log();
   console.log(`🍾 ${brightGreen("Site built into")} ${gray(site.options.dest)}`);
+
+  if (site.options.metrics) {
+    if (options.metrics) {
+      const file = join(Deno.cwd(), options.metrics);
+      await site.metrics.save(file);
+      console.log();
+      console.log(`⏲ ${brightGreen("Metrics data saved in")} ${gray(file)}`);
+      console.log();
+    } else {
+      console.log();
+      console.log(`⏲ Metrics data:`);
+      console.log();
+      site.metrics.print();
+    }
+  }
 
   if (!options.serve) {
     return;
   }
 
   try {
+    // Disable metrics for the watcher
+    site.options.metrics = false;
+
     await server(site);
     const watcher = Deno.watchFs(site.src());
     const changes = new Set();